Build against wordpress-rs trunk now that the stats PRs have landed
Both the per-post stats endpoint and the follow-up that remodelled its
response are merged, so the branch no longer needs to point at an open PR.

Trunk carries that remodel, which changes two things here. The endpoint now
takes a target rather than a raw id, so the home page is addressed as
StatsPostTarget.HomePage instead of post 0 -- translated at the data source,
leaving the rest of the app on a plain post id. And averages are integers,
so Avg. Views Per Day goes through the same formatter and threshold as every
other figure on the screen instead of the Double overload.

A trunk-<sha> pin is still a moving target; this wants a released version
before it merges.

+++ WordPress/src/main/java/org/wordpress/android/ui/newstats/datasource/StatsDataSource.kt
@@ -869,7 +869,7 @@ data class PostViewsYear(
  */
 data class PostViewsYearAverage(
     val year: String,
-    val overall: Double
+    val overall: Long
 )
 
 /**
